Engine relocation batch: flux_engine + cutoff_risk_engine
Two more engines relocated under ADR-018's per-tool layout. Shim
pattern matches the Sprint 753 recon pilot.

flux_engine.py (288 lines) -> services/audit/flux/analysis.py
- Now sits next to services/audit/flux/recon.py (Sprint 753) in the
  same per-tool package
- Updated services/audit/flux/recon.py to import FluxResult from the
  sibling instead of the legacy flux_engine path
- 10 callers continue to work via the shim

cutoff_risk_engine.py (338 lines) -> services/audit/cutoff_risk/analysis.py
- New per-tool package for the ISA 501 cutoff indicator engine
- 2 callers (audit/pipeline.py, tests/test_cutoff_risk.py) work via
  the shim; test imports the underscore-prefixed helpers, so the
  shim re-exports them too with explicit __all__

backend/tests/test_engine_shims.py (new): parametrized contract test
for both relocations. Asserts symbol identity between shim + canonical
path, and that __all__ matches the documented surface. Generalizes
beyond test_recon_engine_shim.py — future relocations add a row to
SHIM_TARGETS rather than a new test file.

lint_domain_relocation.py advisory count: 31 -> 29.

137 affected tests pass (recon, recon_shim, cutoff_risk, audit_flux_routes,
export_routes, export_diagnostics_routes); 25 shim + lint tests pass.
